refactor: extract LaTeX rendering logic into separate renderLatex module
